requirements.txt
fastapi[all]
jose
openai
passlib
pydantic[email]
python-dotenv
python_jose
python_jose
SQLAlchemy
sqlmodel
uvicorn
pymysql
